fix lint/coverage/mypy
