+++ monkeytype/type_metadata/__init__.py
@@ -1,0 +1,2 @@
+from .__base__ import *  # noqa
+from .encode_decode import *  # noqa
